Split lookup store into own file / class
